updated project with latest changes

+++ backend/routes.js
@@ -5,9 +5,9 @@ import authMiddleware from "./authMiddleware.js"
 const router = Router()
 
 // get all todos
-router.get("/todos", async (req, res) => {
+router.get("/todos", authMiddleware, async (req, res) => {
     try {
-        const result = await pool.query("SELECT * FROM todos ORDER BY created_at DESC")
+        const result = await pool.query("SELECT * FROM todos WHERE user_id = $1 ORDER BY created_at DESC", [request.userId])
         res.json(result.rows)
     } catch (error) {
         res.status(500).json({ error: error.message })
@@ -15,10 +15,10 @@ router.get("/todos", async (req, res) => {
 })
 
 // create a todo
-router.post("/todos", async (req, res) => {
+router.post("/todos",  authMiddleware, async (req, res) => {
     try {
         const { title } = req.body
-        const result = await pool.query("INSERT INTO todos (title) VALUES ($1) RETURNING *", [title])
+        const result = await pool.query("INSERT INTO todos (title, user_id) VALUES ($1, $2) RETURNING *", [title, req.userId])
         res.json(result.rows[0])
     } catch (error) {
         res.status(500).json({ error: error.message })
@@ -29,7 +29,7 @@ router.post("/todos", async (req, res) => {
 router.put("/todos/:id", async (req, res) => {
     try {
         const { id } = req.params
-        const result = await pool.query("UPDATE todos SET is_completed = NOT is_completed WHERE id = $1 RETURNING *", [id])
+        const result = await pool.query("UPDATE todos SET is_completed = NOT is_completed WHERE id = $1 AND user_id = $2 RETURNING *", [id, req.userId])
         res.json(result.rows[0])
     } catch (error) {
         res.status(500).json({ error: error.message })
@@ -37,11 +37,11 @@ router.put("/todos/:id", async (req, res) => {
 }) 
 
 // update title
-router.put("/todos/:id/title", async (req, res) => {
+router.put("/todos/:id/title", authMiddleware, async (req, res) => {
     try {
         const { id } = req.params
         const { title } = req.body
-        const result = await pool.query("UPDATE todos SET title = $1 WHERE id = $2 RETURNING *", [title, id])
+        const result = await pool.query("UPDATE todos SET title = $1 WHERE id = $2 AND user_id = $3 RETURNING *", [title, id, req.userId])
         res.json(result.rows[0])
     } catch (error) {
         res.status(500).json({ error: error.message })
@@ -49,10 +49,10 @@ router.put("/todos/:id/title", async (req, res) => {
 })
 
 // delete a todo
-router.delete("/todos/:id", async (req, res) => {
+router.delete("/todos/:id", authMiddleware, async (req, res) => {
     try {
         const { id } = req.params
-        await pool.query("DELETE FROM todos WHERE id = $1", [id])
+        await pool.query("DELETE FROM todos WHERE id = $1 AND user_id = $2", [id, req.userId])
         res.json({ message: "Todo deleted" })
     } catch (error) {
         res.status(500).json({ error: error.message })
